CORREZIONE BUG

+++ ThermostatFraschetta.h
@@ -4,7 +4,7 @@
 #include "VirtualCycleFraschetta.h"
 enum ThermostatType{Heating,Cooling};
 class ThermostatF:RelayF,VirtualCycleF{
-protected: ThermostatType TypeT; uint8_t PinRelay; byte TemperatureThresholdS; bool State;
+protected: ThermostatType TypeT; uint8_t PinRelay; int8_t TemperatureThresholdS; bool State;
  public:
  ThermostatF(){}
  ThermostatF(uint8_t PinRelay,DigitalType ReleType=NormalLogic,ThermostatType TypeT=Heating):RelayF(PinRelay,ReleType),
@@ -13,18 +13,18 @@ protected: ThermostatType TypeT; uint8_t PinRelay; byte TemperatureThresholdS; b
  using RelayF::ReadDigitalStatus;
  void Begin(){
   RelayF::Begin();
-  if(EEPROM.read(PinRelay)>49){EEPROM.update(PinRelay,-20);}
+  if(EEPROM.read(PinRelay)>49){EEPROM.update(PinRelay,int8_t(-20));}
   TemperatureThresholdS=EEPROM.read(PinRelay);
   if(TemperatureThresholdS<0){TemperatureThresholdS=-TemperatureThresholdS;}else{State=1;}
  }
- void SetTemperatureThreshold(byte TemperatureThreshold){
+ void SetTemperatureThreshold(uint8_t TemperatureThreshold){
   TemperatureThresholdS=TemperatureThreshold;
   if(!State){TemperatureThreshold=-TemperatureThreshold;}
   EEPROM.update(PinRelay,TemperatureThreshold);
  }
  void TurnOn(){State=1; SetTemperatureThreshold(TemperatureThresholdS);}
  void TurnOff(){State=0; SetTemperatureThreshold(TemperatureThresholdS);}
- void ThermostatFunction(byte CurrentTemperature){
+ void ThermostatFunction(uint8_t CurrentTemperature){
   if(nTick()>0 && State){
    if(TypeT==Heating){if(CurrentTemperature<TemperatureThresholdS){RelayF::TurnOn();} else{RelayF::TurnOff();}}
    else{if(CurrentTemperature>TemperatureThresholdS){RelayF::TurnOn();} else {RelayF::TurnOff();}}
